refactor(p4-4/3): draw guards → campaignDrawGuards.js; the copy-pasted closesAt/terms 422s live once
